refactor: split admin page into components
